requirements.txt
Flask
icecream
pymongo
python-dotenv
Werkzeug
aiogram